// src/modules/inventory/reports/accessories-store/accessories-receive-status-by-challan-no-report/accessories-receive-status-by-challan-no-report-type.ts
export interface IAccessoriesReceiveStatusByChallanNoReport {
  ITEM_NO: string;
  BUYER_ID: number;
  STYLE_ID: number;
  COLOR_ID: number;
  SIZE_ID: number;
  PO_ID: number;
  PURCHASE_ORDER_NO: string;
  RECEIVE_DATE: Date;
  CHALLAN_DATE: Date;
  RECEIVE_QTY: number;
  COMPANY_NO: number;
  ITEM_NAME: string;
  BUYER_NAME: string;
  STYLENO: string;
  COLORNAME: string;
  SIZENAME: string;
  PONO: string;
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
  SUB_PO: string;
  MRR_NO: string;
  DESCRIPTION_1: string;
  DESCRIPTION_2: string;
  UOM: string;
  ALLOCATION_NO: string;
  RET_NO: string;
}